restructure code and separate out stats utils and mounter utils
